refactor: 패키지구조 수정에 따른 테스트코드 import 변경
