+++ src/test/java/com/hhpl/concertreserve/domain/waitingqueue/WaitingQueueTest.java
@@ -1,6 +1,9 @@
 package com.hhpl.concertreserve.domain.waitingqueue;
 
 import com.hhpl.concertreserve.domain.error.BusinessException;
+import com.hhpl.concertreserve.domain.waitingqueue.model.WaitingQueue;
+import com.hhpl.concertreserve.domain.waitingqueue.model.WaitingQueueInfo;
+import com.hhpl.concertreserve.domain.waitingqueue.type.WaitingQueueStatus;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.test.util.ReflectionTestUtils;
@@ -9,8 +12,8 @@ import java.time.LocalDateTime;
 
 import static com.hhpl.concertreserve.domain.error.BusinessExceptionCode.QUEUE_IS_EXPIRED;
 import static com.hhpl.concertreserve.domain.error.BusinessExceptionCode.QUEUE_IS_INACTIVE;
-import static com.hhpl.concertreserve.domain.waitingqueue.WaitingQueueStatus.EXPIRED;
-import static com.hhpl.concertreserve.domain.waitingqueue.WaitingQueueStatus.INACTIVE;
+import static com.hhpl.concertreserve.domain.waitingqueue.type.WaitingQueueStatus.EXPIRED;
+import static com.hhpl.concertreserve.domain.waitingqueue.type.WaitingQueueStatus.INACTIVE;
 import static org.junit.jupiter.api.Assertions.*;
 
 class WaitingQueueTest {
